server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
supabase==2.9.1
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.1
mangum==0.17.0
